Fix Unique Constraint error When Assigning DAO to 'No DAO' Contribution
Previously, the logic would check for existing guild activity types when a user attempted to assign a DAO to a contribution that didn't have an associated DAO.

If no guild activity types were found, the resolver would attempt to create a new one. However, this leads to a Unique Constraint Failure error if the user activity type already existed.